Provide a helper for building a NavTree
I'm not sure how to use it yet, though. It has to be generated at build
time, but `getStaticProps` can't be used in the main `_app` because of
limitations in NextJS. The only solution I can see is to repeatedly use
it in every page we make, which isn't DRY and is prone to error.

I'm looking into possibly generating a JSON file manually during build
(ie, without using `getStaticProps`) that the `_app` page would be able
to read, but not sure that will work. Also, it would probably be prone
to error as well.