Provide a helper for building a NavTree (#69)
I decided to go the more NextJS-friendly route, and update all pages to
build the nav tree in `getStaticProps`. The containing `_app` page has a
build-time check to verify this happens (if we forget to build the nav
in one of the pages, the build will break).

I also updated some links to point to pages that exist.